requirements.txt
discord.py==1.2.3
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
psycopg2==2.8.3
gunicorn==19.9.0
beautifulsoup4==4.7.1
schedule==0.6.0
pytz==2019.2